+++ samples/dotnet/clinician-org-sample/src/Pages/Error.cshtml.cs
@@ -13,7 +13,7 @@ namespace Lenus.Samples.ClinicianOrg.Pages
     [IgnoreAntiforgeryToken]
     public class ErrorModel : PageModel
     {
-        public string RequestId { get; set; }
+        public string RequestId { get; set; } = string.Empty;
 
         public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
