+++ samples/dotnet/clinician-org-sample/src/Services/LenusOrganisationalMembership.cs
@@ -1,0 +1,10 @@
+using System;
+
+namespace Lenus.Samples.ClinicianOrg.Services
+{
+    public class LenusOrganisationalMembership
+    {
+        public Guid Id { get; set; }
+        public string? Name { get; set; }
+    }
+}
