+++ samples/dotnet/clinician-org-sample/src/appsettings.json
@@ -9,14 +9,15 @@
   "AllowedHosts": "*",
   "Lenus": {
     "OpenIdConnect": {
-      "Authority": "https://identity.sbx.ggc.lenushealth.dev",
-      "ClientId": "philo.clinicalorgsample"
+      "Authority": "https://<identity-host>",
+      "ClientId": "",
+      "ClientSecret" : ""
     },
     "Agency": {
-      "BaseApiUri": "https://api.sbx.ggc.lenushealth.dev/agency/v1"
+      "BaseApiUri": "https://<api-host>/agency/v1"
     },
     "Organisations": {
-      "BaseApiUri": "https://api.sbx.ggc.lenushealth.dev/organizations/v1"
+      "BaseApiUri": "https://<api-host>/organizations/v1"
     }
   }
 }
